feat(runtime): share tracing infrastructure
Move generic command execution and source metadata into reusable packages while keeping profiler-specific orchestration focused.
