+++ internal/utils/executil/exec.go
@@ -15,15 +15,154 @@
 package executil
 
 import (
+	"bytes"
+	"context"
+	"errors"
 	"fmt"
 	"os"
+	"os/exec"
 	"path/filepath"
 	"runtime"
+	"strings"
+	"syscall"
 
+	"huatuo-bamai/internal/log"
 	"huatuo-bamai/internal/procfs"
 
 	"golang.org/x/sys/unix"
 )
+
+type CmdResult struct {
+	Pid     int
+	Cmd     string
+	Stdout  []byte
+	Stderr  []byte
+	Success bool
+	CmdErr  error
+}
+
+// ExecCmd executes one command, captures its output, and terminates its
+// process group when the context is canceled.
+func ExecCmd(ctx context.Context, pid int, binPath string, args ...string) CmdResult {
+	cmdArgs := formatCmd(binPath, args)
+	log.Debugf("executing command: %s", cmdArgs)
+
+	cmd := exec.CommandContext(ctx, binPath, args...)
+	cmd.Env = os.Environ()
+	cmd.Stdin = os.Stdin
+	cmd.SysProcAttr = &syscall.SysProcAttr{Setpgid: true}
+
+	var stdoutBuf, stderrBuf bytes.Buffer
+	cmd.Stdout = &stdoutBuf
+	cmd.Stderr = &stderrBuf
+
+	if err := cmd.Start(); err != nil {
+		return CmdResult{
+			Pid:     pid,
+			Cmd:     cmdArgs,
+			Stderr:  stderrBuf.Bytes(),
+			Success: false,
+			CmdErr:  err,
+		}
+	}
+
+	done := make(chan error, 1)
+	go func() {
+		done <- cmd.Wait()
+	}()
+
+	select {
+	case <-ctx.Done():
+		if err := syscall.Kill(-cmd.Process.Pid, syscall.SIGTERM); err != nil && !errors.Is(err, syscall.ESRCH) {
+			log.Warnf("kill process group %d: %v", cmd.Process.Pid, err)
+		}
+		<-done
+		cmdErr := ctx.Err()
+		log.Debugf("command stopped: command=%q error=%v", cmdArgs, cmdErr)
+		return CmdResult{
+			Pid:     pid,
+			Cmd:     cmdArgs,
+			Stdout:  stdoutBuf.Bytes(),
+			Stderr:  stderrBuf.Bytes(),
+			Success: false,
+			CmdErr:  cmdErr,
+		}
+	case err := <-done:
+		return CmdResult{
+			Pid:     pid,
+			Cmd:     cmdArgs,
+			Stdout:  stdoutBuf.Bytes(),
+			Stderr:  stderrBuf.Bytes(),
+			Success: err == nil,
+			CmdErr:  err,
+		}
+	}
+}
+
+func formatCmd(binPath string, args []string) string {
+	return binPath + " " + strings.Join(args, " ")
+}
+
+const maxCommandOutputInError = 4096
+
+type commandResultError struct {
+	pid    int
+	cmd    string
+	stdout string
+	stderr string
+	err    error
+}
+
+func (e *commandResultError) Error() string {
+	var message strings.Builder
+	fmt.Fprintf(&message, "command %q failed for pid %d", e.cmd, e.pid)
+	if e.err != nil {
+		fmt.Fprintf(&message, ": %v", e.err)
+	}
+	if e.stderr != "" {
+		fmt.Fprintf(&message, "; stderr=%q", e.stderr)
+	}
+	if e.stdout != "" {
+		fmt.Fprintf(&message, "; stdout=%q", e.stdout)
+	}
+
+	return message.String()
+}
+
+func (e *commandResultError) Unwrap() error {
+	return e.err
+}
+
+func commandOutputForError(output []byte) string {
+	trimmed := strings.TrimSpace(string(output))
+	if len(trimmed) <= maxCommandOutputInError {
+		return trimmed
+	}
+
+	return trimmed[:maxCommandOutputInError] + "... (truncated)"
+}
+
+// VerifyResults reports every failed command with its process and captured
+// diagnostics so callers can identify the failing tool invocation.
+func VerifyResults(results []CmdResult) error {
+	failures := make([]error, 0)
+	for _, result := range results {
+		if result.Success {
+			continue
+		}
+
+		log.Debugf("command failed: command=%q pid=%d error=%v", result.Cmd, result.Pid, result.CmdErr)
+		failures = append(failures, &commandResultError{
+			pid:    result.Pid,
+			cmd:    result.Cmd,
+			stdout: commandOutputForError(result.Stdout),
+			stderr: commandOutputForError(result.Stderr),
+			err:    result.CmdErr,
+		})
+	}
+
+	return errors.Join(failures...)
+}
 
 func RunningDir() (string, error) {
 	exePath, err := os.Executable()
